+++ bridgetown-core/lib/bridgetown-core/concerns/convertible.rb
