+++ bridgetown-core/lib/bridgetown-core/drops/site_drop.rb
@@ -7,7 +7,7 @@ module Bridgetown
 
       mutable false
 
-      def_delegator  :@obj, :site_data, :data
+      def_delegator  :@obj, :data
       def_delegators :@obj, :time, :pages, :static_files, :tags, :categories
 
       private def_delegator :@obj, :config, :fallback_data
